requirements.txt
anthropic==0.84.0
anyio==4.12.1
asyncpg==0.31.0
discord.py==2.7.0
fastapi==0.115.0
httpx==0.28.1
python-dotenv==1.2.1
redis==7.2.1
uvicorn==0.34.0
pydantic==2.12.5
